add transac query process to components
